% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailwind_cli.R
\name{install_tailwindcss_cli}
\alias{install_tailwindcss_cli}
\title{Installs the tailwindcss CLI}
\usage{
install_tailwindcss_cli(
  basedir = NULL,
  overwrite = FALSE,
  version = "latest",
  verbose = FALSE
)
}
\arguments{
\item{basedir}{directory where you want the tailwindcss executable to be
installed}

\item{overwrite}{if existing installations should be overwritten}

\item{version}{the version to install, default is latest}

\item{verbose}{if the version etc should be reported}
}
\value{
the path to the tailwindcss CLI invisibly
}
\description{
This will download the tailwindcss standalone CLI to the folder
\code{basedir}.
}
\details{
This will download the tailwindcss standalone CLI to the folder
\code{basedir}.
See \href{https://tailwindcss.com/blog/standalone-cli}{here} for details on the
standalone CLI. This saves you from having to install \code{node.js}.

On the mac, after installing the CLI, you need to make sure that the file
is executable to run it. The easiest way to do so, is to \code{cd} to your
\code{basedir} from the terminal and then run \verb{chmod +x tailwindcss} on Mac.
}
\examples{
if (interactive()) {
  install_tailwindcss_cli()
}
}
\seealso{
\link{compile_tailwindcss}
}
