% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailwind_cli.R
\name{tailwindcss_compile}
\alias{tailwindcss_compile}
\title{Starts the tailwindcss CLI}
\usage{
tailwindcss_compile(
  infile,
  outfile,
  watch = FALSE,
  tailwindcss = NULL,
  verbose = FALSE
)
}
\arguments{
\item{infile}{the tailwindcss css file (eg containing the \verb{@tailwind} directives)}

\item{outfile}{the target css file, where tailwind will write the css to}

\item{watch}{if the files should be continuously monitored (versus only
compile the css once), default is False}

\item{tailwindcss}{name and path to the executable}

\item{verbose}{print information}
}
\value{
the outfile invisibly
}
\description{
See also \href{https://tailwindcss.com/blog/standalone-cli}{docs}
}
\examples{
if (interactive()) {
  temp <- tempdir()
  owd <- setwd(temp)

  infile <- "custom.css"
  writeLines(
    c("@tailwind base;","@tailwind components;", "@tailwind utilities;"),
    infile)
  outfile <- "out.css"

  file.copy(system.file("examples", "01-Old_Faithful", "app.R", package = "shiny.tailwind"),
            "app.R", overwrite = TRUE)

  tailwindcss_compile(infile, outfile)
  cat(paste(readLines(outfile)[1:20], collapse = "\n"))

  setwd(owd)
}
}
