% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twCheckboxInput.R
\name{twCheckboxInput}
\alias{twCheckboxInput}
\title{Title}
\usage{
twCheckboxInput(
  inputId,
  label = NULL,
  value = FALSE,
  width = NULL,
  disabled = FALSE,
  container_class = NULL,
  label_class = NULL,
  input_class = NULL
)
}
\arguments{
\item{input_class}{}
}
\description{
Title
}
\examples{

shiny::checkboxInput("id", "label", value = FALSE)
twCheckboxInput("id", "label", value = TRUE,
                container_class = "CONTAINER", label_class = "LABEL",
                input_class = "INPUT")

if (interactive()) {
library(shiny)
# basic example
shinyApp(
    ui = fluidPage(
        use_tailwind(),
        twCheckboxInput(
          "chk", "Check me!", value = TRUE,
          container_class = "runded-tl-lg bg-gray-200 m-4 p-2",
          label_class = "font-serif text-red-500",
          input_class = "checked:focus:bg-red-200 hover:bg-red-500 bg-red-200"
        ),
        verbatimTextOutput("out")
    ),
    server = function(input, output) {
        output$out <- renderText({input$chk})
    }
)
}

}
