% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-twSelect.R
\name{twSelectInput}
\alias{twSelectInput}
\title{Wrapper around shiny::selectInput but allowing for more classes}
\usage{
twSelectInput(
  inputId,
  label = NULL,
  choices,
  selected = NULL,
  multiple = FALSE,
  container_class = NULL,
  label_class = NULL,
  select_class = NULL
)
}
\description{
Note selectize is turned off (for now)
}
