% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_flowbite.R
\name{use_flowbite}
\alias{use_flowbite}
\title{Allows you to use flowbite components}
\usage{
use_flowbite(version = "1.4.7")
}
\arguments{
\item{version}{the version of flowbite to use, default is 1.4.7}
}
\value{
the required HTML-HEAD tags
}
\description{
See also: \url{https://flowbite.com/} and \url{https://flowbite.com/#components}
}
\examples{
asd
}
