% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_preline.R
\name{use_preline}
\alias{use_preline}
\title{Allows you to use 'preline' components}
\usage{
use_preline(version = "3.0.0", ...)
}
\arguments{
\item{version}{the version of 'preline' to use, default is 3.0.0}

\item{...}{further arguments passed to \code{\link[=use_tailwind]{use_tailwind()}}}
}
\value{
the required HTML-head tags to use 'preline' as \code{shiny.tag}
}
\description{
See also: \url{https://preline.co/} and \url{https://preline.co/docs/index.html}
}
