% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailwind_cli.R
\name{install_tailwindcss_cli}
\alias{install_tailwindcss_cli}
\title{Installs the tailwindcss CLI}
\usage{
install_tailwindcss_cli(overwrite = FALSE, version = "latest", verbose = FALSE)
}
\arguments{
\item{overwrite}{if existing installations should be overwritten}

\item{version}{the version to install, default is latest}

\item{verbose}{if the version etc should be reported}
}
\value{
the path to the tailwindcss CLI invisibly
}
\description{
Installs the tailwindcss CLI
}
\examples{
if (interactive()) {
  install_tailwindcss_cli()
}
}
\seealso{
\link{tailwindcss_compile}
}
