% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-twText.R
\name{twTextInput}
\alias{twTextInput}
\title{Wrapper around shiny::textInput but allowing for more classes}
\usage{
twTextInput(
  inputId,
  label = NULL,
  value = NULL,
  placeholder = NULL,
  type = "text",
  container_class = NULL,
  label_class = NULL,
  input_class = NULL
)
}
\description{
Wrapper around shiny::textInput but allowing for more classes
}
